Fixed Mac memory module to compile.

+++ src/memory/mac7m.c
@@ -1,9 +1,9 @@
 
-#include "../memory.h"
+#include "../dstypes.h"
 
 #include <Multiverse.h>
 
-MEMORY_HANDLE memory_alloc( uint16_t sz, uint16_t count ) {
+MEMORY_HANDLE memory_alloc( uint32_t sz, uint32_t count ) {
    /* TODO: Detect overflow. */
    return NewHandleClear( (sz * count) );
 }
@@ -12,20 +12,20 @@ void memory_free( MEMORY_HANDLE handle ) {
    DisposeHandle( handle );
 }
 
-uint16_t memory_sz( MEMORY_HANDLE handle ) {
+uint32_t memory_sz( MEMORY_HANDLE handle ) {
    return GetHandleSize( handle );
 }
 
-uint16_t memory_resize( MEMORY_HANDLE handle, uint16_t sz ) {
+uint32_t memory_resize( MEMORY_HANDLE handle, uint32_t sz ) {
    SetHandleSize( handle, sz );
    return sz;
 }
 
-void memory_copy_ptr( void* dest, const void* src, uint16_t sz ) {
+void memory_copy_ptr( void* dest, const void* src, uint32_t sz ) {
    BlockMove( src, dest, sz );
 }
 
-void memory_zero_ptr( void* ptr, uint16_t sz ) {
+void memory_zero_ptr( void* ptr, uint32_t sz ) {
    memset( ptr, 0, sz );
 }
